requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
cssselect2==0.8.0
Django==5.2.3
django-cors-headers==4.7.0
django-crum==0.7.9
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
fonttools==4.58.4
html5lib==1.1
idna==3.10
lxml==6.0.0
mysqlclient==2.2.7
numpy==2.3.1
openpyxl==3.1.5
oscrypto==1.3.0
pandas==2.3.1
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyJWT==2.10.1
PyMySQL==1.1.1
pypdf==5.8.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.3
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
weasyprint==65.1
webencodings==0.5.1
xhtml2pdf==0.2.17
xlrd==2.0.2
zopfli==0.2.3.post1
